fix: inline the capability sandbox into march_runtime.c
A separate runtime/march_sandbox.c broke `march repl` and 33 tests.
march_spawn_main calls march_sandbox_install unconditionally, and every
harness linking the runtime keeps its own source list -- bin/main.ml,
four test/dune rules, test_helpers.ml, and the REPL's JIT .so builder --
so a new translation unit must be added to all of them or the link fails
with an undefined symbol.

Fixed by deleting the file and inlining the function behind
#ifdef MARCH_CAP_PROFILE, which removes the class of breakage rather
than patching six lists. A new runtime .c file is a multi-site change; a
new function in an existing one is not.

Verified after: march repl works, run_eval 256/256, cap_strip 4/4,
--cap-sandbox still embeds a module-specific profile (pure binary has no
allow network, net binary does), default builds carry no profile.
